Make actually compile
I have maximum creativity at night and minimum competence.

I'd had a thinko at the initial calloc and it got worse from there.

+++ tests/consistency/native.h
@@ -8,12 +8,12 @@
 
 // Memcmp returns -1,0,1. 0 for equality
 
-static inline int v6_equal (const unsigned char *p1,
+static inline size_t v6_equal (const unsigned char *p1,
                             const unsigned char *p2) {
        return memcmp(p1,p2,16) == 0;
 }
 
-static inline int v6_nequal (const unsigned char *p1,
+static inline size_t v6_nequal (const unsigned char *p1,
                             const unsigned char *p2) {
        return memcmp(p1,p2,16) ;
